+++ visualization/app/codeCharta/ui/layoutSelection/layoutSelection.module.ts
@@ -1,7 +1,0 @@
-import "../../state/state.module"
-import angular from "angular"
-import { layoutSelectionComponent } from "./layoutSelection.component"
-
-angular
-	.module("app.codeCharta.ui.layoutSelection", ["app.codeCharta.state"])
-	.component(layoutSelectionComponent.selector, layoutSelectionComponent)