+++ visualization/app/codeCharta/ui/metricChooser/metricChooser.component.ts
@@ -1,124 +1,37 @@
 import "./metricChooser.component.scss"
-import { IRootScopeService } from "angular"
-import { NodeMetricData } from "../../codeCharta.model"
-import { StoreService } from "../../state/store.service"
-import { setAreaMetric } from "../../state/store/dynamicSettings/areaMetric/areaMetric.actions"
-import { setHeightMetric } from "../../state/store/dynamicSettings/heightMetric/heightMetric.actions"
-import { setDistributionMetric } from "../../state/store/dynamicSettings/distributionMetric/distributionMetric.actions"
-import { setColorMetric } from "../../state/store/dynamicSettings/colorMetric/colorMetric.actions"
-import { AreaMetricService, AreaMetricSubscriber } from "../../state/store/dynamicSettings/areaMetric/areaMetric.service"
-import { HeightMetricService, HeightMetricSubscriber } from "../../state/store/dynamicSettings/heightMetric/heightMetric.service"
-import { ColorMetricService, ColorMetricSubscriber } from "../../state/store/dynamicSettings/colorMetric/colorMetric.service"
-import {
-	DistributionMetricService,
-	DistributionMetricSubscriber
-} from "../../state/store/dynamicSettings/distributionMetric/distributionMetric.service"
-import { NodeMetricDataService, NodeMetricDataSubscriber } from "../../state/store/metricData/nodeMetricData/nodeMetricData.service"
+import { Component, ElementRef, Inject, Input, OnInit, ViewChild } from "@angular/core"
+import { Store } from "../../state/angular-redux/store"
+import { nodeMetricDataSelector } from "../../state/selectors/accumulatedData/metricData/nodeMetricData.selector"
+import { Observable } from "rxjs"
+import { EdgeMetricData, NodeMetricData } from "../../codeCharta.model"
+import { edgeMetricDataSelector } from "../../state/selectors/accumulatedData/metricData/edgeMetricData.selector"
 
-export class MetricChooserController
-	implements NodeMetricDataSubscriber, AreaMetricSubscriber, HeightMetricSubscriber, ColorMetricSubscriber, DistributionMetricSubscriber
-{
-	private originalMetricData: NodeMetricData[]
+type MetricChooserType = "node" | "edge"
 
-	private _viewModel: {
-		metricData: NodeMetricData[]
-		areaMetric: string
-		colorMetric: string
-		heightMetric: string
-		distributionMetric: string
-		searchTerm: string
-	} = {
-		metricData: [],
-		areaMetric: null,
-		colorMetric: null,
-		heightMetric: null,
-		distributionMetric: null,
-		searchTerm: ""
+@Component({
+	selector: "cc-metric-chooser",
+	template: require("./metricChooser.component.html")
+})
+export class MetricChooserComponent implements OnInit {
+	@Input() selectedMetricName: string
+	@Input() searchPlaceholder: string
+	@Input() handleMetricChanged: (newSelectedMetricName: string) => void
+	@Input() type: MetricChooserType = "node"
+	@ViewChild("searchTermInput") searchTermInput: ElementRef<HTMLInputElement>
+	searchTerm = ""
+	metricData$: Observable<NodeMetricData[] | EdgeMetricData[]>
+
+	constructor(@Inject(Store) private store: Store) {}
+
+	ngOnInit(): void {
+		this.metricData$ = this.store.select(this.type === "node" ? nodeMetricDataSelector : edgeMetricDataSelector)
 	}
 
-	constructor(private $rootScope: IRootScopeService, private storeService: StoreService) {
-		"ngInject"
-		AreaMetricService.subscribe(this.$rootScope, this)
-		HeightMetricService.subscribe(this.$rootScope, this)
-		ColorMetricService.subscribe(this.$rootScope, this)
-		DistributionMetricService.subscribe(this.$rootScope, this)
-		NodeMetricDataService.subscribe(this.$rootScope, this)
+	handleOpenedChanged(opened: boolean) {
+		if (opened) {
+			this.searchTermInput.nativeElement.focus()
+		} else {
+			this.searchTerm = ""
+		}
 	}
-
-	onAreaMetricChanged(areaMetric: string) {
-		this._viewModel.areaMetric = areaMetric
-	}
-
-	onHeightMetricChanged(heightMetric: string) {
-		this._viewModel.heightMetric = heightMetric
-	}
-
-	onColorMetricChanged(colorMetric: string) {
-		this._viewModel.colorMetric = colorMetric
-	}
-
-	onDistributionMetricChanged(distributionMetric: string) {
-		this._viewModel.distributionMetric = distributionMetric
-	}
-
-	onNodeMetricDataChanged(nodeMetricData: NodeMetricData[]) {
-		this._viewModel.metricData = nodeMetricData
-		this.originalMetricData = nodeMetricData
-	}
-
-	filterMetricData() {
-		const searchTerm = this._viewModel.searchTerm.toLowerCase()
-		this._viewModel.metricData = this.originalMetricData.filter(({ name }) => name.toLowerCase().includes(searchTerm))
-	}
-
-	clearSearchTerm() {
-		this._viewModel.searchTerm = ""
-		this._viewModel.metricData = this.originalMetricData
-	}
-
-	focusInputField(idName: string) {
-		setTimeout(() => {
-			document.getElementById(`${idName}-selector`).focus()
-		}, 200)
-	}
-
-	applySettingsAreaMetric() {
-		this.storeService.dispatch(setAreaMetric(this._viewModel.areaMetric))
-	}
-
-	applySettingsColorMetric() {
-		this.storeService.dispatch(setColorMetric(this._viewModel.colorMetric))
-	}
-
-	applySettingsHeightMetric() {
-		this.storeService.dispatch(setHeightMetric(this._viewModel.heightMetric))
-	}
-
-	applySettingsDistributionMetric() {
-		this.storeService.dispatch(setDistributionMetric(this._viewModel.distributionMetric))
-	}
-}
-
-export const areaMetricChooserComponent = {
-	selector: "areaMetricChooserComponent",
-	template: require("./metricChooser.area.component.html"),
-	controller: MetricChooserController
-}
-
-export const heightMetricChooserComponent = {
-	selector: "heightMetricChooserComponent",
-	template: require("./metricChooser.height.component.html"),
-	controller: MetricChooserController
-}
-
-export const colorMetricChooserComponent = {
-	selector: "colorMetricChooserComponent",
-	template: require("./metricChooser.color.component.html"),
-	controller: MetricChooserController
-}
-
-export const distribitionMetricChooserComponent = {
-	selector: "distributionMetricChooserComponent",
-	template: require("./metricChooser.distribution.component.html"),
-	controller: MetricChooserController
 }