+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.e2e.ts
@@ -1,8 +1,8 @@
 import { goto } from "../../../puppeteer.helper"
 import { RibbonBarPageObject } from "./ribbonBar.po"
 import { SearchPanelPageObject } from "../searchPanel/searchPanel.po"
-import { AreaSettingsPanelPageObject } from "../areaSettingsPanel/areaSettingsPanel.po"
 import { UploadFileButtonPageObject } from "../toolBar/uploadFilesButton/uploadFilesButton.po"
+import { AreaSettingsPanelPageObject } from "./areaSettingsPanel.po"
 
 //Commented out flaky test
 
@@ -46,40 +46,40 @@ describe("RibbonBar", () => {
 		it("height-metric card", async () => {
 			const panel = "height-metric"
 
-			let isHeightSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			let isHeightSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(panel)
 			expect(isHeightSettingsPanelOpen).toBeTruthy()
 
-			isHeightSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			isHeightSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(panel)
 			expect(isHeightSettingsPanelOpen).toBeFalsy()
 		})
 
 		it("area-metric card", async () => {
 			const panel = "area-metric"
 
-			let isAreaSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			let isAreaSettingsPanelOpen = await ribbonBar.togglePanel(panel, "cc-area-settings-panel")
 			expect(isAreaSettingsPanelOpen).toBeTruthy()
 
-			isAreaSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			isAreaSettingsPanelOpen = await ribbonBar.togglePanel(panel, "cc-area-settings-panel")
 			expect(isAreaSettingsPanelOpen).toBeFalsy()
 		})
 
 		it("color-metric card", async () => {
 			const panel = "color-metric"
 
-			let isColorSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			let isColorSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(panel)
 			expect(isColorSettingsPanelOpen).toBeTruthy()
 
-			isColorSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			isColorSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(panel)
 			expect(isColorSettingsPanelOpen).toBeFalsy()
 		})
 
 		it("edge-metric card", async () => {
 			const panel = "edge-metric"
 
-			let isEdgeSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			let isEdgeSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(panel)
 			expect(isEdgeSettingsPanelOpen).toBeTruthy()
 
-			isEdgeSettingsPanelOpen = await ribbonBar.togglePanel(panel)
+			isEdgeSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(panel)
 			expect(isEdgeSettingsPanelOpen).toBeFalsy()
 		})
 	})
@@ -88,30 +88,30 @@ describe("RibbonBar", () => {
 		const areaPanel = "area-metric"
 		const edgePanel = "edge-metric"
 
-		let isAreaSettingsPanelOpen = await ribbonBar.togglePanel(areaPanel)
+		let isAreaSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(areaPanel)
 		expect(isAreaSettingsPanelOpen).toBeTruthy()
 
 		const isSearchPanelOpen = await searchPanel.toggle()
 		expect(isSearchPanelOpen).toBeTruthy()
-		expect(await ribbonBar.isPanelOpen(areaPanel)).toBeFalsy()
+		expect(await ribbonBar.isPanelOpenAngularJS(areaPanel)).toBeFalsy()
 
-		isAreaSettingsPanelOpen = await ribbonBar.togglePanel(areaPanel)
+		isAreaSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(areaPanel)
 		expect(isAreaSettingsPanelOpen).toBeTruthy()
 		expect(await searchPanel.isOpen()).toBeFalsy()
 
-		const isEdgeSettingsPanelOpen = await ribbonBar.togglePanel(edgePanel)
+		const isEdgeSettingsPanelOpen = await ribbonBar.togglePanelAngularJS(edgePanel)
 		expect(isEdgeSettingsPanelOpen).toBeTruthy()
-		expect(await ribbonBar.isPanelOpen(areaPanel)).toBeFalsy()
+		expect(await ribbonBar.isPanelOpenAngularJS(areaPanel)).toBeFalsy()
 	})
 
 	it("should open a section and keep it open after clicking a button inside it", async () => {
 		const areaPanel = "area-metric"
 
-		await ribbonBar.togglePanel(areaPanel)
+		await ribbonBar.togglePanelAngularJS(areaPanel)
 		expect(AreaSettingsPanelPageObject.isDefaultMarginEnabled()).toBeTruthy()
 		expect(await AreaSettingsPanelPageObject.toggleDefaultMargin()).toBeFalsy()
 
-		expect(await ribbonBar.isPanelOpen(areaPanel)).toBeTruthy()
+		expect(await ribbonBar.isPanelOpenAngularJS(areaPanel)).toBeTruthy()
 	})
 
 	it("should not find edge-metric-panel when file has no edge metrics", async () => {