+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.po.ts
@@ -1,17 +1,29 @@
 import { clickButtonOnPageElement } from "../../../puppeteer.helper"
 
+/** When the ribbon bar is migrated we should replace this helper and the related e2e test through @testing-library/angular tests */
 export class RibbonBarPageObject {
 	private EXPANDED = "expanded"
 	private EXPANDED_REMOVE = "expanded-remove"
 
-	async isPanelOpen(selector: string) {
+	async isPanelOpenAngularJS(selector: string) {
 		await page.waitForSelector(`#${selector}-card`)
 		const classNames = await page.$eval(`#${selector}-card`, element => element["className"])
 		return classNames.includes(this.EXPANDED) && !classNames.includes(this.EXPANDED_REMOVE)
 	}
 
-	async togglePanel(selector: string) {
-		const wasOpen = await this.isPanelOpen(selector)
+	async isPanelOpen(elementName: string) {
+		await page.waitForFunction(elementName => document.querySelector(elementName), {}, elementName)
+		return page.$eval(elementName, element => !element.classList.contains("hidden"))
+	}
+
+	async togglePanelAngularJS(selector: string) {
+		const wasOpen = await this.isPanelOpenAngularJS(selector)
+		await clickButtonOnPageElement(`#${selector}-card .section .section-title`)
+		return !wasOpen
+	}
+
+	async togglePanel(selector: string, elementName: string) {
+		const wasOpen = await this.isPanelOpen(elementName)
 		await clickButtonOnPageElement(`#${selector}-card .section .section-title`)
 
 		await (wasOpen ? page.waitForSelector(`#${selector}-card`) : page.waitForSelector(`#${selector}-card.${this.EXPANDED}`))